fixed bug with smg and pistol not doing damage after switching guns

+++ Rapid Fire Interactive's FPS Mid-Term/Assets/Scripts/playerMovement.cs
@@ -887,7 +887,7 @@ public class playerMovement : MonoBehaviour, IDamage
 
     void changeGun() {
         float damTemp = guns[gunPos].damage * damageUpgradeMod;
-        damage = (int)damTemp;
+        damage = damTemp;
         bulletDistance = guns[gunPos].bulletDist;
         fireRate = getCurGun().fireRate;
         isSniper = getCurGun().isSniper;
